fix: Update Channels and Operator Image in Bundle
This updates the operator bundle to specify the channels it can be
added to within an OLM catalog, and to the latest operator image
published on Konflux. The operator uses the "latest" channel for minor
version updates, and "openshift-builds-1.y" for patch version updates.
The Makefile was also updated to set the VERSION variable so we can
properly test operator updates within a catalog.

The Makefile IMAGE_TAG_BASE was updated to match the default push
target for the bundle and any associated images on Konflux. This value
is used when running the `make bundle` Makefile target.

The Makefile IMG value was also updated to use a valid image + tag.
This will facilitate testing of the operator bundle (especially within
an FBC catalog). This value is used when running `make bundle`, as this
process sets the operator image via `kustomize edit` before rendering
the operator bundle.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>