feat: Test Operator with File-Based Catalog
Use OPM to render an OLM file-based catalog for testing. The base
manifests are stored in the config/catalog directory, which contain an
upgrade graph for Builds for OpenShift in its current state. This also
includes a CatalogSource that can be deployed on OpenShift via
Kustomize. Makefile targets and additional documentation added to help
developers test the operator end to end, the way a cluster-admin will
manage the operator.

fix line ending