BUILD-993: Add Builds to Supported OCP Catalog Indexes
This updates our bundle lables and annotations to specify the versions
of OCP the operator can be deployed on (4.12-4.17). As part of this
work, the operator CSV was updated to set the minimum k8s version to
1.25.0. A few minor fixes to CSV were also added.

Note that the labels on `bundle.Dockerfile` and the annotations in
`bundle/manifests/metadata/annotations.yaml` need to be synced
manually. This is due to a feature gap in operator-sdk's
`generate bundle` command.

Co-authored-by: Sayan Biswas <sayan-biswas@live.com>
Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>