chore: Bump operator base images
Switch to using the UBI go-toolset image for compiling the operator.
Starting with go 1.21, the compiler is capable of compiling go code
with newer toolchains [1]. Therefore, we no longer need to wait for
RHEL to publish go-toolset 1.22 images in order to compile 1.22 code.

There is some risk that by setting the GOTOOLCHAIN env var to a value
other than "local" (the default in the go-toolset image), we are unable
to build the operator in a hemetic environment.

[1] https://go.dev/blog/toolchain