Boostrap OperatorHub e2e Test
Updating the e2e test scaffolding from operator-sdk to deploy Builds
for OpenShift using OperatorHub. The updated e2e uses the "standard"
install method from the Red Hat operators catalog, whose YAML manifest
is embedded into the `setup.go` test framework code. Kubernetes objects
are deployed using the default controller-runtime client and
`unstructured.Unstructured` objects with TypeMeta set.

The `BeforeSuite` phase checks if Builds for OpenShift has been
installed via OLM, and if not creates the required `Subscription`. It
then waits for the generated OLM `ClusterServiceVersion` object to
exist and reach the `Succeeded` phase. Subsequent tests should
implement their own `BeforeEach` logic to ensure the specific system
under test is ready (usually an operand). An example is provided, where
the operator deployment is checked for readiness before verification of
the `OpenShiftBuild` CRD is run.

Teardown logic for the operator is not provided due to the nested
dependencies of the Builds for OpenShift operator. This can be
addressed in a future test enhancement.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>