Install Test for OpenShift Builds
Initial install e2e test for OpenShift Builds. This test assumes
operator has been subscribed with OLM. Once the operator deployment
reports itself "Ready", the test then verifies the default settings for
the OpenShiftBuild instance and the status of operands.

Ginkgo labels are used to distinguish which feature sets are being
tested. In the future, these can be used to run subsets of the e2e test
suite. The following labels are used as a starting point:

- e2e, operator: indicate this is an e2e test of the whole operator
- shipwright: tests of Shipwright features
- builds: tests of build features (will likely be paired with
  shipwright label)
- shared-resources: tests of the Shared Resource CSI Driver

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>