Setup Async Reconciler Tests
Convert the existing "unit" test suite to an integration test suite
that runs the controllers in a separate goroutine. The setup is based
on the example documented in the Kubebuilder v4 book [1]. The test
begins with an empty instance of the OpenShiftBuild object, which
uncovered a few gaps in our controller-driven defaulting logic. A few
bits of the operator controller code were refactored to accomodate
this use case.

Our flavor of the ShipwrightBuild controller/reconciler is not included
in this test suite. Likewise, the bootstrapping tests were commented
out, since this collides with the controller code running in a separate
goroutine. Future test updates should exercise all controllers and use
unit-style tests for the bootstrap logic.

In order to reconcile SharedResources, additional CRDs had to be
installed into our EnvTest environment. This currently includes the
Prometheus Operator CRDs for v0.78.1. Future CRDs that need to be added
should be downloaded to the `test/integration/crd` directory.

[1] https://book.kubebuilder.io/cronjob-tutorial/writing-tests

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>