Fix bundle generation with relatedImages
Changes:
Right now the `relatedImages` section is not correctly generated because the operand images used are declared in the environment variable and hence hte `--use-image-digest` only extracts and updates the images present in the CSV in deployment spec `image` field only.
This fix adds a YQ transformation to copy `relatedImages` from base CSV to bundle CSV.

Also, the version is fixed, pointing to the correct version.