[BUILD-1432]: Mounts root filesystem as read-only in Shared Resources containers and add E2E test
This commit mounts the root filesystem as read-only in the Shared Resources containers. The `readOnlyRootFilesystem: true` setting is applied to the DaemonSet and Webhook containers. This is a security best practice that prevents the container's internal files from being modified at runtime. To ensure the driver still functions correctly, the `workingDir` for the `hostpath` container is set to the writable `/run/csi-data-dir`. This allows the driver to handle temporary files without compromising the security of its root filesystem. The new E2E test verifies this entire workflow by successfully sharing both Secrets and ConfigMaps, keeping the shared resources' functionality intact.