[BUILD-1432] : Mounts root filesystem as read-only in Shared Resources containers and add E2E test
This commit mounts the root filesystem as read-only in the Shared Resources containers. The  setting is applied to the DaemonSet and Webhook containers. This is a security best practice that prevents the container's internal files from being modified at runtime. This change sets /run/csi-data-dir—a hostPath mounted from the host's writable tmpfs, specifically intended for the CSI driver's data operations—as the default starting directory for the driver's main process. This allows the driver to handle temporary files without compromising the security of its root filesystem. The new E2E test verifies this entire workflow by successfully sharing both Secrets and ConfigMaps, keeping the shared resources' functionality intact.

Signed-off-by: Anchita Borah <borahanchita@gmail.com>